Merge fix for PduParser.java - Part 1
It looks like there was a merge issue in PduParser.java that duplicated
some code when the gingerbread code was merged.  The change in
gingerbread essentially put the entire PduPart.P_CONTENT_DISPOSITION in
an if statement, and set it to positive for all devices.

This fixes MMS problems on the Nexus One, and possibly other phones.

Change-Id: I47b0f17327f906055e1d358226b5db0ddfb2241f